requirements.txt
Flask==1.1.2
gunicorn==20.0.4
mysql-connector==2.2.9
